MDEV-26632: GTID master switch when slave position is filtered on new master
If an intermediate slave S1 has replication filters enabled, its
@@gtid_slave_pos may contain a GTID that is filtered and doesn't propagate
to lower-level slaves with S1 as master.

If then later S1 is demoted to a slave, it may attempt to connect to the
filtered position. This is normally disallowed in --gtid-strict-mode. But if
--gtid-ignore-duplicates is enabled, we should allow it, as in this case we
can trust the GTID sequence numbers between different server ids. So we can
know that the next GTID is the right one for the filtered slave GTID
position.

This allows advanced users to use replication filtering in topologies like
this and still run with --gtid-strict-mode enabled.

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>