MDEV-32830 I. refactor XA binlogging for better integration with BGC/replication/recovery
This commit is the part I in the series of three that addresses MDEV-31949.

Here the task is to improve upon MDEV-742 design of the XA binlogging.
Specifically, when binlog is ON, to execute XA-{prepare, commit,
rollback} via MYSQL_BIN_LOG::run_ordered_commit(), so that at the end of
a respective operation the XA transaction state is fully shaped
as far as (parallel) replication/recovery is concerned.
In particular, its XID is available for next use, incl for
XA-"complete" by another slave parallel XA-commit/rollback or XA-start of
the new same XID transaction.

Notable changes:

sql/handler.cc
  - ha_prepare() now is ensured to execute binlog_hton::prepare() as
    the first branch to prepare, which is also consistent with
    the MDEV-21117's commit_one_phase_2() policy.
    This measure enforces the write-to-binlog first rule for XA:s.
    The engine(s)'s prepare is realized via an extended
    MYSQL_BIN_LOG::run_ordered_commit()
  - read-only cases are handled more uniformly with extending
    has_binlog_hton() that registers the binlog branch at ha_prepare()
    when necessary
  - ha_rollback_trans() executes binlog_hton::rollback() as first
    branch too
  - ditto to the external completion of XA via ha_commit_or_rollback_by_xid().
sql/log.cc
  - binlog_commit,rollback() simplified/clarified on the condition of an empty
    statement/trx
  - MYSQL_BIN_LOG::run_commit_ordered() is extended to cover
    XA-{prepare, commit, rollback} respective actions in Engines
  - XID unlogging by XA is converted to employ the standard
    binlog-checkpoint mechanism (cache_mngr->using_xa= TRUE)
sql/xa.cc
   - trans_xa_commit,rollback() are refactored to invoke a common function
     for external (connection) completion
   - adapting the above two and slave_applier_reset_xa_trans() to
     possible (normal to slave) XID release, inside run_ordered_commit().