MDEV-32830 I. refactor XA binlogging for better integration with BGC/replication/recovery
This commit is the part I in the series of three that addresses MDEV-31949.

This part improves upon MDEV-742 design of the XA binlogging.
When binlog is ON, to execute XA-prepare first in engines to write
the xa-prepare replication event last.
MYSQL_BIN_LOG::run_ordered_commit() is made to execute XA commit and
rollback in engines, as well as
it completes with the XA transaction state, so that its
XID has become available for next use, e.g by another slave parallel
worker, that e.g could be waiting for its order to commit a being
prepared XID transaction.

Notable changes:

sql/handler.cc
  - ha_prepare() now is ensured to execute binlog_hton::prepare() as
    the last branch to prepare; recovery concerns are addressed
    separately (not by MDEV-21469, but rather along the plan of bug#76233)
  - read-only cases are handled more uniformly with extending
    has_binlog_hton() that registers the binlog branch at ha_prepare()
    when necessary
  - ha_rollback_trans() executes binlog_hton::rollback() as first
    branch, as a part of the 76233 planned recovery;
  - ditto to the external completion of XA via ha_commit_or_rollback_by_xid().
sql/log.cc
  - binlog_commit,rollback() simplified/clarified on the condition of an empty
    statement/trx
  - MYSQL_BIN_LOG::run_commit_ordered() is extended to cover
    XA-{commit, rollback} respective actions in Engines
  - XID unlogging by XA is converted to employ the standard
    binlog-checkpoint mechanism (cache_mngr->using_xa= TRUE)
sql/xa.cc
   - trans_xa_commit,rollback() are refactored to invoke a common function
     for external (connection) completion
   - adapting the above two and slave_applier_reset_xa_trans() to
     possible (normal to slave) XID release, inside run_ordered_commit().