MDEV-32686: crash information to include operating system information
To make crash information resolve easier to a distribution
lets include this information in the crash output.

Its reasonable to support non-Windows filesystems could support
the path /etc/os-release as evidence of FreeBSD and Solaris
support exist (but not a Windows path so exclude this).

3 lines of /etc/os-release is sufficient to identify os/distro and
version in most cases (sampled from
https://github.com/which-distro/os-release). Samples otherwise
can be long, and include SUPPORT_URL/BUG_REPORT_URL to distros
which in the case of crashes, aren't as easily resolvable by
distros. Those normally aren't in the first 3 lines.

We place this information before the core information, example output:

    The manual page at https://mariadb.com/kb/en/how-to-produce-a-full-stack-trace-for-mariadbd/ contains
    information that should help you find out what is causing the crash.

    Distribution information:
    NAME="Fedora Linux"
    VERSION="39 (Workstation Edition)"
    ID=fedora

    Writing a core file...

To support Apple/FreeBSD kernel information its 'len' variable was
renamed to avoid conflicts and the shared 'buff' is used.

Small corrections that should have been fixed in MDEV-30613.
* open() doesn't take a MY_F argument but its previous my_open did.
* casting to uchar isn't needed as it was previous my_read which
  required a case.

Also close files as soon the read is made.

Being a crash hander we don't output any error messages on failing
to open/read/close any of these files as it could be confusing
to the user bug report. They might assume that failing to get
this information is related to the cause of the signal which it
isn't.