MDEV-32551: "Read semi-sync reply magic number error" warnings on master
The test and the related commit message comes from Brandon Nesterenko

A test to show how to induce the ‘Read semi-sync reply magic number
error’ message on a primary. In short, if semi-sync is turned on
during the hand-shake process between a primary and replica, but
later a user negates the rpl_semi_sync_slave_enabled variable while
the replica’s IO thread is running; if the io thread exits, the
replica can skip a necessary call to kill_connection() in
repl_semisync_slave.slave_stop() due to its reliance on a global
variable. Then, the replica will send a COM_QUIT packet to the
primary on an active semi-sync connection, causing the magic number
error.

The test in this patch exits the IO thread by forcing an error;
though note a call to STOP SLAVE could also do this, but it ends up
needing more synchronization. That is, the STOP SLAVE command also
tries to kill the vio of the replica, which makes a race with the IO
thread to try and send the COM_QUIT before this happens (which would
need more debug_sync to get around). See THD::awake_no_mutex for
details as to the killing of the replica’s vio.

Changes done:
- Some timings changed slight with startup of slave which caused
  rpl_binlog_dump_slave_gtid_state_info.text to fail as it checked the
  error log file before the slave had started properly. Fixed by adding
  wait_for_pattern_in_file.inc that allows waiting for the pattern to
  appear in the log file.
- The rpl_semi_sync_slave_enabled variable is now a default value for
  when semisync is started. The variable does not anymore affect
  semisync if it is already running. This fixes the original reported bug.
  Internally we now use repl_semisync_slave.get_slave_enabled() instead of
  rpl_semi_sync_slave_enabled. To check if semisync is active on should
  check the @@rpl_semi_sync_slave_status variable (as before).
- Changed the default value for rpl-semi-sync-master-wait-no-slave to 0 as
  the old one did not make much sense as default. The main difference is
  that now we do not wait for semisync ack if there are no slave threads.
  In the old code, we waited once, which did not bring any notable benefits
  except slower startup of master configured for using semisync.
- Removed some duplicate semisync notes from the error log.
- We do not call semisync.switch_off() if there are no active slaves.
  Instead we check in Repl_semi_sync_master::commit_trx() if there are
  no active threads. This simplices a lot of code.
- Changed assert() to DBUG_ASSERT() to ensure that the DBUG log is flushed
  in case of asserts.
- Removed Repl_semi_sync_master::check_and_switch() as it is not needed
  anymore.
- Remove rpl_semi_sync_slave_status as it is not needed anymore.
- Removed rpl_semi_sync_slave_enabled  as it is not needed anymore.
  Repl_semi_sync_slave::get_slave_enabled() contains the active status.
- Added checking that we do not add a slave twice with
  Ack_receiver::add_slave(). This could happen with old code.
- Ensure that when we call Ack_receiver::remove_slave() that the slave
  is removed from the listener before function returns.
- Call listener.listen_on_sockets() outside of mutex for better performance
  and less contested mutex.
- Ensure that listening is ignoring newly added slaves when checking for
  responses.
- kill_zombie_dump_threads() now does killing of dump threads properly.
- Added an error if kill_zombie_dump_threads() fails.

Co-author: Brandon Nesterenko <brandon.nesterenko@mariadb.com>