MDEV-32808: using sql_service within encryption key management plugin segfaults
There is an assumption in the way the SQL service initializes threads
that assumes a thread already exists. For a key management plugin
this isn't the case. As such we perform the initializations on
fil_crypt_thread() so all the SQL layer functionality doesn't
segfault on the lack of a THD.

Create a THD option for the innodb crypt thread.

Co-author: Daniel Black <daniel@mariadb.org>