MDEV-32852 assert in xa-rollback branch of MDEV-32830 run_commit_ordered
is replaced.
XA in either XA_ACTIVE or XA_IDLE state can
be ended by connection close which in combination with
a non-safe to rollback property can lead into the assert branch as well.

A proper assert is placed to check the fact of a forceful xa ending
with connection close.