MDEV-27771 [ERROR] mysqld: Incorrect information in file: './test/t1.frm' on "ERROR 1918 Encountered illegal value ... when converting to latin1"
The server could write to the FRM file bad 'GENERATED ALWAYS AS'
expressions which later do not parse at the FRM open time.
This led to messages written to the server error log.

Fix:

To parse safely at open time, simple string literals must convert to the
table level CHARACTER SET, because @@collation_connection is set to the
table level CHARACTER SET/COLLATE during parsing of the virtual column
expressions.

Additional fix:

If the 'GENERATED ALWAYS AS' expression of a character string column
is a literal, this literal must safely convert to the column character set.
Otherwise this value is useless and will raise errors during DML statements.
This change makes the behavior of these two definitions symmetric:
- GENERATES ALWAYS AS ('non-convertable string literal')
- DEFAULT 'non-convertable string literal'

Note, more complex cases with a string literal inside is not detected yet, e.g.:
- GENERATES ALWAYS AS (func('non-convertable string literal'))